requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.57
botocore==1.35.57
certifi==2024.8.30
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
jmespath==1.0.1
numpy==2.1.3
openai==1.54.3
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
